+++ src/components/Menu/TripletNotesMenu.tsx
@@ -1,24 +1,51 @@
 import StickingsMenu from './StickingsMenu/StickingsMenu';
 import TripletsStaff from '../Staff/TripletsStaff/TripletsStaff';
 import { tripletPermutations } from '../../lib/utils/permutations';
+import { useState } from 'react';
+import './Menu.css';
 
-interface Props {
-  onFormChange: (beatName: string, children: string) => void;
-  onReset: () => void;
-  selectedStickings: { [key: string]: string };
-}
+// interface Props {
+//   onFormChange: (beatName: string, children: string) => void;
+//   onReset: () => void;
+//   selectedStickings: { [key: string]: string };
+// }
 
-function TripletNotesMenu({ onFormChange, onReset, selectedStickings }: Props) {
+function TripletNotesMenu() {
+  const [selectedStickings, setSelectedStickings] = useState<{
+    [key: string]: string;
+  }>({});
+
+  const handleFormChange = (beatName: string, children: string) => {
+    if (beatName === 'row') {
+      setSelectedStickings({
+        'beat-1': children,
+        'beat-2': children,
+        'beat-3': children,
+        'beat-4': children,
+      });
+      return;
+    }
+    setSelectedStickings({
+      ...selectedStickings,
+      [beatName]: children,
+    });
+  };
+
+  const handleResetClick = () => {
+    setSelectedStickings({});
+  };
+
+  console.log(selectedStickings);
   return (
-    <>
+    <div className="menu">
       <TripletsStaff selectedStickings={selectedStickings} />
       <StickingsMenu
         permutations={tripletPermutations}
-        onFormChange={onFormChange}
-        onReset={onReset}
+        onFormChange={handleFormChange}
+        onReset={handleResetClick}
         selectedStickings={selectedStickings}
       />
-    </>
+    </div>
   );
 }
